+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a bug where guards would raise exceptions instead of just being false
 
+### Added
+
+- Added support for the OTP `socket` interface.
+
 ## [0.6.0-alpha.1] - 2023-10-09
 
 ### Added